Fix 32-bit overflow in startBrake() ACCELERATING branch
steps_remaining = step_count * profile.accel / profile.decel computes
the multiplication in 32-bit signed long before dividing, so it can
overflow with legal short accel/decel values (e.g. accel=32767,
decel=32000 at step_count=70000 yields steps_remaining=-62539 instead
of the correct 71677). A negative steps_remaining is treated as
STOPPED by getCurrentState(), so nextAction() stops issuing step
pulses immediately instead of ramping down, defeating startBrake()'s
purpose of a controlled deceleration.

Cast to float before the multiply, matching the same overflow-avoidance
idiom already used elsewhere in this file (line 187, and the sibling
steps_to_cruise/steps_to_brake computation in startMove()).

Verified by compiling the affected source against a minimal Arduino.h
host mock and driving the public API (setRPM/setSpeedProfile/startMove/
nextAction/startBrake) to reproduce the overflow with real, legal
parameters, then confirming the patched result exactly matches the
64-bit ground truth. Also ran 6 non-overflowing regression scenarios
(rpm 60-2000, microsteps 1-64, accel/decel 100-32767) with byte-identical
output before/after the change.

Agent: claude-sonnet-5 (Claude Code)

+++ src/BasicStepperDriver.cpp
@@ -222,7 +222,9 @@ void BasicStepperDriver::startBrake(void){
         break;
 
     case ACCELERATING:
-        steps_remaining = step_count * profile.accel / profile.decel;
+        // compare in float to avoid 32-bit overflow of step_count * profile.accel
+        // with high microstep/rpm/accel combinations (same pattern as startMove())
+        steps_remaining = (float)step_count * profile.accel / profile.decel;
         break;
 
     default:
